2026/click/requirements.txt
python-wayland==1.0.0
